requirements.txt
streamlit==1.29.0
boto3==1.34.0
pandas==2.1.4
redis==5.0.1
python-dotenv==1.0.0
plotly==5.17.0
numpy==1.24.3

#                                                                                                                                                                 
# To activate this environment, use                                                                                                                               
#                                                                                                                                                                 
#     $ conda activate D:\Projects\DataQuery\venvquery
#
# To deactivate an active environment, use
#
#     $ conda deactivate
#  pip install -r requirements.txt